[Git2Git] Git Train: Merge of building/rs_onecore_dep_aci_oss/170502-1700 into official/rs_onecore_dep_aci_oss Retrieved from https://microsoft.visualstudio.com os OS official/rs_onecore_dep_aci_oss ff0f40448513c17201a643780cf538d63f45e41b

Cherry picked from Pull request 424143.

Related work items: #11877850

+++ src/renderer/wddmcon/WddmConRenderer.cpp
@@ -256,7 +256,7 @@ HRESULT WddmConEngine::PaintBufferLine(PCWCHAR const pwsLine, size_t const cchLi
 
     if (_hWddmConCtx != INVALID_HANDLE_VALUE)
     {
-        for (size_t i = 0 ; i < cchLine && i < _displayWidth ; i++)
+        for (size_t i = 0 ; i < cchLine && i < (size_t)_displayWidth ; i++)
         {
             OldChar = &_displayState[coord.Y]->Old[coord.X + i];
             NewChar = &_displayState[coord.Y]->New[coord.X + i];
